# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Rest RCS Express",
    "description": "API Rest RCS Express",
    "contact": {
      "name": "rcsxpress",
      "email": "operaciones@rcsexpress.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Ambiente de Desarrollo"
    },
    {
      "url": "https://scen.rcsexpress.com/api/v1/",
      "description": "Ambiente de Producción"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/zonas": {
      "get": {
        "tags": ["Zonas"],
        "summary": "Obtener todas las zonas",
        "description": "API para Obtener todas las zonas del sistema",
        "parameters": [
          {
            "name": "agencia",
            "in": "header",
            "description": "Codigo de la Agencia",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Zonas"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Zonas"],
        "summary": "Crear Zonas",
        "description": "API para Crear una nueva Zona",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Zonas"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Zonas"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/zonas/{id}": {
      "get": {
        "tags": ["Zonas"],
        "summary": "Obtener Zona por su Id",
        "description": "API para Retornar una Zona por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Zona",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Zonas"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Zonas"],
        "summary": "Actualizar Zonas",
        "description": "API para Actualizar una Zona por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Zona",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Zonas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Zonas"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Zonas"],
        "summary": "Borrar Zonas",
        "description": "API para Borrar una Zona por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Zona",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Zonas/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Zonas": {
        "description": "Modelo de Zonas",
        "required": ["id", "nb_zona", "tipo_zona", "cod_agencia"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "nb_zona": {
            "type": "string"
          },
          "tipo_zona": {
            "type": "string"
          },
          "cod_agencia": {
            "type": "number"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}